images/keycloak/pyproject.toml
[tool.poetry]
name = "keycloak"
version = "0.1.6"
description = ""
authors = ["Andreas Altenkirch <andreas.altenkirch@solenix.ch>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-keycloak = "^3.9.1"
pyjwt = "^2.8.0"

[tool.poetry.group.test.dependencies]
pytest = "^7"
coverage = {extras = ["toml"], version = "^7"}

[tool.poetry.group.dev.dependencies]
isort = "^5.12"
pylint = "^2.17"
black = "^23.7"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
py_version = 311
profile = "black"
combine_as_imports = true
src_paths = ["data_uploader", "tests"]

[tool.pylint]
max-line-length = 120

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
    "integration",
]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "if __name__ == .__main__.:",
    ]

[tool.coverage.xml]
output = "coverage-report/coverage.xml"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
keycloak_login = "src.main:main"
